관리자

+++ Kmarket/src/main/java/kr/co/kmarket/controller/admin/Admin_NoticeSelectDeleteController.java
@@ -3,7 +3,6 @@ package kr.co.kmarket.controller.admin;
 import java.io.IOException;
 import java.util.Arrays;
 
-import javax.servlet.RequestDispatcher;
 import javax.servlet.ServletException;
 import javax.servlet.annotation.WebServlet;
 import javax.servlet.http.HttpServlet;
@@ -15,21 +14,26 @@ import org.slf4j.LoggerFactory;
 
 import kr.co.kmarket.service.CsService;
 
-@WebServlet("/admin/faq/delete.do")
-public class Admin_FaqDeleteController extends HttpServlet {
+@WebServlet("/admin/notice/seletDelete.do")
+public class Admin_NoticeSelectDeleteController extends HttpServlet {
 
     private static final long serialVersionUID = 1232344352L;
-    private Logger logger = LoggerFactory.getLogger(Admin_FaqDeleteController.class);
+    private Logger logger = LoggerFactory.getLogger(Admin_NoticeSelectDeleteController.class);
     private CsService service = CsService.INSTANCE;
 
     @Override
     protected void doGet(HttpServletRequest req, HttpServletResponse resp) throws ServletException, IOException {
+        String[] checkBoxArr = req.getParameterValues("checkBoxArr");
+        logger.debug("checkBoxArr : " + Arrays.toString(checkBoxArr));
         
-    	String no = req.getParameter("no");
-       
-	    service.deleteBoard(no);
-	    
-		RequestDispatcher dispatcher = req.getRequestDispatcher("/admin/faq/list.jsp?group=2");
-		dispatcher.forward(req, resp);
+        if (checkBoxArr != null && checkBoxArr.length != 0) {
+  
+	            for (int i = 0; i < checkBoxArr.length; i += 1) {
+	            String no = checkBoxArr[i];            
+	            logger.debug("no : " + no);
+	            
+	            service.deleteBoard(no);
+            }
+        }
     }
 }
